p2g_website/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# add JWT login + protect admin endpoints
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
